move the page/book/bundle maker helpers
