+++ server/src/model/book.spec.ts
@@ -1,13 +1,12 @@
 import { BookValidationKind } from './book'
-import { expectErrors, first, loadSuccess, makeBundle } from './spec-helpers'
-import { pageMaker } from './page.spec'
+import { bookMaker, BookMakerTocNode, expectErrors, first, loadSuccess, makeBundle, pageMaker } from './spec-helpers'
 
 describe('Book validations', () => {
   it(BookValidationKind.DUPLICATE_CHAPTER_TITLE.title, () => {
     const bundle = makeBundle()
     const book = first(loadSuccess(bundle).books)
     const chapterTitle = 'Kinematics'
-    const toc: TocNode[] = [
+    const toc: BookMakerTocNode[] = [
       { title: chapterTitle, children: [] },
       { title: chapterTitle, children: [] }
     ]
@@ -24,7 +23,7 @@ describe('Book validations', () => {
   it(BookValidationKind.DUPLICATE_PAGE.title, () => {
     const bundle = makeBundle()
     const book = first(loadSuccess(bundle).books)
-    const toc: TocNode[] = [
+    const toc: BookMakerTocNode[] = [
       { title: 'Chapter 1', children: ['m00001'] },
       { title: 'Chapter 2', children: ['m00001'] }
     ]
@@ -34,50 +33,3 @@ describe('Book validations', () => {
     expectErrors(book, [BookValidationKind.DUPLICATE_PAGE, BookValidationKind.DUPLICATE_PAGE])
   })
 })
-
-type TocNode = {
-  title: string
-  children: TocNode[]
-} | string
-interface BookMakerInfo {
-  title?: string
-  slug?: string
-  uuid?: string
-  language?: string
-  licenseUrl?: string
-  toc?: TocNode[]
-}
-export function bookMaker(info: BookMakerInfo) {
-  const i = {
-    title: info.title ?? 'test collection',
-    slug: info.slug ?? 'slug1',
-    langauge: info.language ?? 'xxyyzz',
-    licenseUrl: info.licenseUrl ?? 'http://creativecommons.org/licenses/by/4.0/',
-    uuid: info.uuid ?? '00000000-0000-4000-0000-000000000000',
-    toc: info.toc ?? []
-  }
-  return `<col:collection xmlns:col="http://cnx.rice.edu/collxml" xmlns:md="http://cnx.rice.edu/mdml" xmlns="http://cnx.rice.edu/collxml">
-    <col:metadata>
-      <md:title>${i.title}</md:title>
-      <md:slug>${i.slug}</md:slug>
-      <md:uuid>${i.uuid}</md:uuid>
-      <md:language>${i.langauge}</md:language>
-      <md:license url="${i.licenseUrl}"/>
-    </col:metadata>
-    <col:content>
-        ${i.toc.map(tocToString).join('\n')}
-    </col:content>
-</col:collection>`
-}
-function tocToString(node: TocNode): string {
-  if (typeof node === 'string') {
-    return `<col:module document="${node}" />`
-  } else {
-    return `<col:subcollection>
-        <md:title>${node.title}</md:title>
-        <col:content>
-            ${node.children.map(tocToString).join('\n')}
-        </col:content>
-    </col:subcollection>`
-  }
-}
